pyproject.toml
[tool.poetry]
name = "contactfinder"
version = "0.1.0"
description = "Find email domain & employee contact."
authors = ["Hojayfa Rahman <hojayfarahman@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
langchain = "^0.3.15"
langchain-openai = "^0.3.3"
langchain-google-genai = "^2.1.5"
langchain-community = "^0.3.27"
beautifulsoup4 = "^4.13.4"
requests = "^2.32.4"
langchainhub = "^0.1.21"
dnspython = "^2.7.0"
google-genai = "^1.25.0"
pydantic = "^2.11.7"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
factory-boy = "^3.3.3"